Cp7 user authorization (#6)
* User Authorization: complete

* Topic show view: minor changes to layout

* Added Simplecov and authorized Test user to edit Topics

+++ app/controllers/bookmarks_controller.rb
@@ -1,24 +1,27 @@
 class BookmarksController < ApplicationController
-  before_action :authenticate_user!
+  before_action :authenticate_user!, except: [:index, :about]
+  rescue_from Pundit::NotAuthorizedError, with: :user_not_authorized
 
-  # GET     ../bookmarks/1
+  # GET ../bookmarks/1
   def show
     @bookmark = Bookmark.find(params[:id])
   end
 
-  # GET     ../bookmarks/new
+  # GET ../bookmarks/new
   def new
     @topic = Topic.find(params[:topic_id])
     @bookmark = @topic.bookmarks.new
   end
 
-  # GET      ../bookmarks/1/edit
+  # GET ../bookmarks/1/edit
   def edit
+    @user = current_user
     @bookmark = Bookmark.find(params[:id])
     @topic = Topic.find(params[:topic_id])
+    authorize @bookmark
   end
 
-  # POST      ../bookmarks
+  # POST ../bookmarks
   def create
     @topic = Topic.find(params[:topic_id])
     @bookmark = @topic.bookmarks.build(bookmark_params)
@@ -33,24 +36,26 @@ class BookmarksController < ApplicationController
     end
   end
 
-  # PATCH(PUT)      ../bookmarks/1
+  # PATCH(PUT) ../bookmarks/1
   def update
     @bookmark = Bookmark.find(params[:id])
     @bookmark.update_attributes(bookmark_params)
 
     if @bookmark.save
-      flash[:notice] = "Your new bookmark has been created!"
+      flash[:notice] = "Your bookmark has been updated!"
       redirect_to @bookmark.topic
     else
-      flash.now[:error] = "Pocket was unable to save new bookmark. Please try again."
+      flash.now[:error] = "Pocket was unable to save this bookmark. Please try again."
       render :edit
     end
   end
 
-  # DELETE        ../bookmarks/1
+  # DELETE ../bookmarks/1
   def destroy
+    @user = current_user
     @bookmark = Bookmark.find(params[:id])
     @topic = Topic.find(params[:topic_id])
+    authorize @bookmark
 
     if @bookmark.destroy
       flash[:notice] = "\"#{@bookmark.url}\" has been deleted."
@@ -61,9 +66,15 @@ class BookmarksController < ApplicationController
     end
   end
 
-  # private
+
+  private
 
   def bookmark_params
     params.require(:bookmark).permit(:url)
+  end
+
+  def user_not_authorized
+    flash[:alert] = "Easy, tiger. Only authorized Pocketeers can do that."
+    redirect_to @bookmark.topic
   end
 end
